seperate widget from screen
