+++ assets/json/targets.json
@@ -1,0 +1,11 @@
+[
+  "chest",
+  "back",
+  "lowerBody",
+  "shoulder",
+  "triceps",
+  "biceps",
+  "core",
+  "forearm",
+  "aerobic"
+]
