+++ lib/routes/app_pages.dart
@@ -1,12 +1,12 @@
 import 'package:get/get.dart';
 
 import 'package:sunmi/binding/workout_binding.dart';
-import 'package:sunmi/ui/screen/Setcount.dart';
-import 'package:sunmi/ui/screen/challenge_list_page.dart';
-import 'package:sunmi/ui/screen/workout_select_page.dart';
+import 'package:sunmi/ui/screen/routine/Setcount.dart';
+import 'package:sunmi/ui/screen/challenge/challenge_list_page.dart';
+import 'package:sunmi/ui/screen/routine/workout_select_page.dart';
 import 'package:sunmi/binding/home_binding.dart';
 import '../data/model/workout.dart';
-import 'package:sunmi/ui/screen/challenge_infomation.dart';
+import 'package:sunmi/ui/screen/challenge/challenge_infomation.dart';
 import 'homepage.dart';
 
 part 'app_routes.dart';
