+++ lib/routes/homepage.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:sunmi/controller/home_controller.dart';
-import 'package:sunmi/ui/screen/challenge_date.dart';
-import 'package:sunmi/ui/screen/challenge_infomation.dart';
-import 'package:sunmi/ui/screen/routine_calendar.dart';
-import 'package:sunmi/ui/screen/sns_navigator.dart';
+import 'package:sunmi/ui/screen/challenge/challenge_date.dart';
+import 'package:sunmi/ui/screen/challenge/challenge_infomation.dart';
+import 'package:sunmi/ui/screen/routine/routine_calendar.dart';
+import 'package:sunmi/ui/screen/sns/sns_navigator.dart';
 
-import '../ui/screen/challenge_list_page.dart';
+import '../ui/screen/challenge/challenge_list_page.dart';
 
 
 class HomePage extends StatefulWidget {
